requirements.txt
Flask>=1.1.1
Jinja2>=2.10.3
SQLAlchemy>=1.3.11
gunicorn>=20.0.0
flask_sqlalchemy>=2.0.0
flask_pymongo>=2.3.0
dnspython==2.1.0